Skip whole `test_capi` if `_testcapi` is not present
